paypal_backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
idna==3.7
paypalrestsdk==1.13.1
pycparser==2.22
pyOpenSSL==24.1.0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.2
six==1.16.0
sqlparse==0.5.0
urllib3==2.2.1
